Move launcher handler to launcher.ts
